backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-ai==0.0.12
google-generativeai==0.3.2
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0

# Additional dependencies for Hakka AI integration
urllib3==2.0.7
certifi==2023.11.17

# Web scraping and content processing
aiohttp==3.9.0
beautifulsoup4==4.12.2
feedparser==6.0.10
lxml==4.9.3